- Summary: [Chart usability spec#2.1]Added Navigation bars for subtasks
- Bugzilla Bug (s) Resolved:
N/A

- Description:
1.	Quotes chart usability spec#2.1:
Description: Add a navigation bar for the subtasks in format chart tab. The navigation bar includes two buttons with drop-down list for moving forward or backward the subtask selection separately. If clicking the arrow-down button of the bar, user can see the history list of the subtask selection.
If some history entries are out of date (e.g. Axis subtask won¡¯t exist in pie chart), they will display ¡°#Invalid¡± and do nothing when being clicked in the drop-down list, or be skipped when being clicked in the button.

2.	Refactored task wizard framework. Move TreeCompoundTask from chart UI to framework

- Tests Description: Manual test
- Notes to Build Team: None
- Notes to Developers: None
- Notes to QA: None
- Notes to Documentation: None

- Files Edited:

- Files Added:

- Files Deleted